+++ src/keys-builder/build-keys.ts
@@ -1,7 +1,6 @@
 import { Config, ScopeMap } from '../types';
 import { checkForProblematicUnflatKeys } from '../utils/keys.utils';
 import { mergeDeep } from '../utils/object.utils';
-
 import { extractTemplateKeys } from './template';
 import { extractTSKeys } from './typescript';
 
@@ -14,6 +13,7 @@ export function buildKeys(config: Config) {
     ts.scopeToKeys
   );
   const fileCount = template.fileCount + ts.fileCount;
+  const defaults = template.defaults.concat(ts.defaults);
 
   if (config.unflat) {
     for (const scopeKeys of Object.values(scopeToKeys)) {
@@ -21,5 +21,5 @@ export function buildKeys(config: Config) {
     }
   }
 
-  return { scopeToKeys, fileCount };
+  return { scopeToKeys, defaults, fileCount };
 }