+++ src/keys-builder/template/index.ts
@@ -1,7 +1,11 @@
-import { Config, ExtractionResult } from '../../types';
+import {
+  Config,
+  DefaultLanguageValue,
+  ExtractionResult,
+  ScopeMap,
+} from '../../types';
 import { readFile } from '../../utils/file.utils';
 import { extractKeys } from '../utils/extract-keys';
-
 import { TemplateExtractorConfig } from './types';
 import { templateCommentsExtractor } from './comments.extractor';
 import { directiveExtractor } from './directive.extractor';
@@ -12,10 +16,15 @@ export function extractTemplateKeys(config: Config): ExtractionResult {
   return extractKeys(config, 'html', templateExtractor);
 }
 
-export function templateExtractor(config: TemplateExtractorConfig) {
+export function templateExtractor(config: TemplateExtractorConfig): {
+  scopeMap: ScopeMap;
+  defaults: DefaultLanguageValue[];
+} {
   const { file, scopeToKeys } = config;
   let content = config.content || readFile(file);
-  if (!content.includes('transloco')) return scopeToKeys;
+  if (!content.includes('transloco')) {
+    return { scopeMap: scopeToKeys, defaults: [] };
+  }
 
   const resolvedConfig = { ...config, content };
   pipeExtractor(resolvedConfig);
@@ -23,5 +32,5 @@ export function templateExtractor(config: TemplateExtractorConfig) {
   directiveExtractor(resolvedConfig);
   structuralDirectiveExtractor(resolvedConfig);
 
-  return scopeToKeys;
+  return { scopeMap: scopeToKeys, defaults: [] };
 }