+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file. See [standa
 ## [1.0.0] (UNRELEASED)
 
 ### Changes 
-* Updated marker function to allow a translation value for the default language (set on extraction).
+* Removed "missing value" string as default for translations. 
 
 ### Features 
 * Added cli and settings options for "useMarkerDefaults" and "defaultLanguage" parameters.
 * Added "workdir" parameter to set the working directory
+* Updated marker function to allow a translation value for the default language (set on extraction).
  
 ### Init
 * Forked "jsverse/transloco-keys-manager" on 2024-02-29.